Global System Message Command
This is just code from /jsb , i have only a bit of lines for it and i do not take this whole part as my own code as i did not make it. Yes this is avaliable for everyone, if you'd like you can make it only for admins. Thanks :P